update tfi token and vesting

+++ src/interfaces/IERC677.sol
@@ -4,16 +4,7 @@ pragma solidity 0.8.19;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IERC677 is IERC20 {
-    event Transfer(
-        address indexed from,
-        address indexed to,
-        uint value,
-        bytes data
-    );
+    event Transfer(address indexed from, address indexed to, uint256 value, bytes data);
 
-    function transferAndCall(
-        address _to,
-        uint _value,
-        bytes calldata _data
-    ) external returns (bool success);
+    function transferAndCall(address _to, uint256 _value, bytes calldata _data) external returns (bool success);
 }
